Move Hero & Contact to App and make AddressForm component
