+++ src/components/AddressForm/index.js
@@ -1,0 +1,2 @@
+export * from './AddressForm'
+export { default } from './AddressForm'
