--- src/components/Contact/index.js
@@ -1,2 +1,0 @@
-export * from './Contact'
-export { default } from './Contact'
