+++ src/components/Footer/Footer.js
@@ -56,7 +56,7 @@ function Footer() {
                                 <h3 className="text-base font-medium text-gray-200">
                                     Solutions
                                 </h3>
-                                <ul role="list" className="mt-4 space-y-4">
+                                <ul className="mt-4 space-y-4">
                                     {footerNavigation.solutions.map((item) => (
                                         <li key={item.name}>
                                             <a
@@ -73,7 +73,7 @@ function Footer() {
                                 <h3 className="text-base font-medium text-gray-200">
                                     Support
                                 </h3>
-                                <ul role="list" className="mt-4 space-y-4">
+                                <ul className="mt-4 space-y-4">
                                     {footerNavigation.support.map((item) => (
                                         <li key={item.name}>
                                             <a
@@ -92,7 +92,7 @@ function Footer() {
                                 <h3 className="text-base font-medium text-gray-200">
                                     Company
                                 </h3>
-                                <ul role="list" className="mt-4 space-y-4">
+                                <ul className="mt-4 space-y-4">
                                     {footerNavigation.company.map((item) => (
                                         <li key={item.name}>
                                             <a
@@ -109,7 +109,7 @@ function Footer() {
                                 <h3 className="text-base font-medium text-gray-200">
                                     Legal
                                 </h3>
-                                <ul role="list" className="mt-4 space-y-4">
+                                <ul className="mt-4 space-y-4">
                                     {footerNavigation.legal.map((item) => (
                                         <li key={item.name}>
                                             <a
