--- src/components/Hero/Hero.js
@@ -1,18 +1,0 @@
-function Hero() {
-    return (
-        <div className="bg-gray-50 py-24 sm:py-32">
-            <div className="mx-auto max-w-md pl-4 pr-8 sm:max-w-lg sm:px-6 lg:max-w-7xl lg:px-8">
-                <h1 className="text-center text-4xl font-bold leading-10 tracking-tight text-gray-900 sm:text-5xl sm:leading-none lg:text-6xl">
-                    Rent Estimate API
-                </h1>
-                <p className="mx-auto mt-6 max-w-3xl text-center text-xl leading-normal text-gray-500">
-                    Get instant access to 140+ million real-time property rent
-                    estimates and comparable rental listings with the Realty
-                    Mole rental data API. Available nationwide in the US.
-                </p>
-            </div>
-        </div>
-    )
-}
-
-export default Hero
