--- src/components/Hero/index.js
@@ -1,2 +1,0 @@
-export * from './Hero'
-export { default } from './Hero'
